requirements.txt
pandas
sqlalchemy
langchain
pinecone-client
gradio
openai
python-dotenv
pyyaml
